+++ py_ce_forms_api/assets/__init__.py
@@ -1,1 +1,2 @@
 from .assets import Assets
+from .asset_elt import AssetElt