+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.6]
+ - Add support for AssetElt
+ - Add processing run awaitable to call process task from command line
+ - Add support for assets array
+
 ## [0.1.5]
  - FIX : manage correctly cancelled task in processing
  - FIX : better error messages and exception handling in processing